ripple section updated design, dotted borders in testimonies
